delete stored number case and its test fixed

+++ src/main/java/io/github/busra/mongo/numbers/restapi/service/StoredNumberService.java
@@ -42,6 +42,10 @@ public class StoredNumberService {
     }
 
     public void deleteStoredNumberByValue(int value) {
+        if (!storedNumberDao.existsByValue(value)) {
+            throw new NotFoundException("No stored number found for value: " + value);
+        }
+
         storedNumberDao.deleteByValue(value);
     }
